[TASK] Remove SchedulerTaskLinkBuilder
